+++ Sources/StreamChatSwiftUI/ChatChannel/Composer/MessageComposerViewModel.swift
@@ -689,6 +689,9 @@ open class MessageComposerViewModel: ObservableObject {
             fetchOptions.predicate = predicate
         }
         fetchOptions.sortDescriptors = [NSSortDescriptor(key: "creationDate", ascending: false)]
+        if let maxGalleryAssetsCount = utils.composerConfig.maxGalleryAssetsCount {
+            fetchOptions.fetchLimit = maxGalleryAssetsCount
+        }
         let assets = PHAsset.fetchAssets(with: fetchOptions)
         DispatchQueue.main.asyncAfter(deadline: .now() + 0.25) { [weak self] in
             self?.imageAssets = assets
@@ -872,6 +875,10 @@ open class MessageComposerViewModel: ObservableObject {
         }
         
         return checkAttachmentSize(with: url)
+    }
+    
+    public func updateAddedAssets(_ assets: [AddedAsset]) {
+        addedAssets = assets
     }
     
     private func checkAttachmentSize(with url: URL?) -> Bool {